requirements.txt
scipy>=1.1
numpy>=1.14
PyVCF>=0.6.8
pysam
pandas==0.23.0
scikit-learn>=0.19
matplotlib>=2.2,<3.0
seaborn>=0.8
pybedtools
tqdm